Rename keccak bin to keccak_bench, add command line opts for measuring things
