+++ risc0/circuit/keccak/Cargo.toml
@@ -19,11 +19,13 @@ tracing = { version = "0.1", default-features = false, features = [
 
 [target.'cfg(not(target_os = "zkvm"))'.dependencies]
 bytemuck = "1.20"
+clap = { version = "4.5", features = ["derive"] }
 cfg-if = { version = "1.0", optional = true }
 risc0-circuit-keccak-sys = { workspace = true, optional = true }
 risc0-circuit-recursion = { workspace = true, default-features = false }
 risc0-sys = { workspace = true, optional = true }
 rayon = { version = "1.10", optional = true }
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 zip = { version = "2.2", default-features = false, features = [
   "deflate",
 ], optional = true }
