+++ risc0/circuit/keccak/src/bin/keccak_bench.rs
@@ -1,0 +1,77 @@
+// Copyright 2024 RISC Zero, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use risc0_circuit_keccak::prove::{keccak_prover, KeccakState};
+
+use clap::Parser;
+use std::time::Instant;
+
+/// Runs a RISC-V ELF binary within the RISC Zero ZKVM.
+#[derive(Parser)]
+#[command(about, version, author)]
+struct Cli {
+    #[arg(long, default_value_t = risc0_circuit_keccak::KECCAK_PO2)]
+    /// Circuit PO2
+    po2: usize,
+
+    #[arg(long, default_value_t = 1)]
+    /// Number of proofs to run
+    count: usize,
+
+    #[arg(long)]
+    /// Don't verify the seal
+    skip_verification: bool,
+}
+
+fn main() {
+    tracing_subscriber::fmt()
+        .with_env_filter(tracing_subscriber::filter::EnvFilter::from_default_env())
+        .init();
+
+    let args = Cli::parse();
+
+    let mut state = KeccakState::default();
+    let mut pows = 987654321_u64;
+    for part in state.as_mut_slice() {
+        *part = pows;
+        pows = pows.wrapping_mul(123456789);
+    }
+
+    let po2 = args.po2;
+    let cycles = 1 << po2;
+    let count = cycles / 200; // roughly 200 cycles per keccakf
+    let inputs = vec![state; count];
+
+    let mut tot_time: f64 = 0.0;
+    for i in 0..args.count {
+        let start_time = Instant::now();
+        let prover = keccak_prover().unwrap();
+        let seal = prover.prove(&inputs, po2).unwrap();
+        if !args.skip_verification {
+            prover.verify(&seal).expect("Verification failed");
+        }
+        let run_time = start_time.elapsed().as_secs_f64();
+        println!(
+            "PO2={po2} Run #{i}: {run_time:.3}s, {:.3} cycles/sec",
+            cycles as f64 / run_time
+        );
+        tot_time += run_time;
+    }
+    println!(
+        "{} runs of PO2={po2} completed in {tot_time:.3}s, avg={:.3}s, {:.3} cycles/sec",
+        args.count,
+        tot_time / (args.count as f64),
+        (args.count * cycles) as f64 / tot_time
+    );
+}
